Rosa cleanup automation (#10)
Checks OCM to see what clusters exists, and compares against the AWS IAM roles that match a given pattern. If there are roles that appear to be for a rosa cluster but no corresponding rosa cluster exists in ocm, then a rosa cli command will be used to delete those roles by prefix.